Move get_yml_list to utils
Move the method without to change it.
Adjust the import in various places.
